refactored: Unified file operations and eliminated code duplication across crates
  - Introduced File::open() and File::open_with_cache() methods that auto-detect compression and dictionary from snapshot headers, eliminating repeated header
  reading/parsing boilerplate across CLI, loader, and server code
  - Consolidated compression creation logic into create_compressor() and create_compressor_from_str() helper functions in hexz_core::algo::compression
  - Added Header::read_from(), Header::read_from_backend(), and Header::load_dictionary() helper methods to standardize header parsing
  - Added MasterIndex::read_from() and MasterIndex::read_from_bounded() for consistent index reading
  - Created ops::inspect module with inspect_snapshot() and SnapshotInfo struct to centralize snapshot metadata extraction, replacing duplicated header/index
  parsing in CLI info command and Python inspect function
  - Created ops::sign module (feature-gated) with sign_snapshot() and verify_snapshot() to centralize cryptographic signing logic, moving it from CLI and Python
  interfaces to core
  - Removed unused compute_snapshot_digest() stub and associated tests from hexz_common::sign
  - Moved shared constants (OVERLAY_BLOCK_SIZE, META_ENTRY_SIZE, CDC chunk sizes) to hexz_common::constants to eliminate duplication across modules
  - Updated SnapshotWriter to correctly handle multiple stream additions by continuing logical positions instead of resetting them
  - Fixed FUSE_BLOCK_SIZE type consistency (now u64 throughout)
  - Updated feature flags: signing now properly propagates from core to loader via hexz-core/signing feature
  - Removed sha2 dependency from loader Cargo.toml (now handled by core signing feature)

+++ crates/fuse/src/vfs/attr.rs
@@ -130,7 +130,7 @@ pub const PERM_FILE: u16 = 0o644;
 /// blocks = file_size.div_ceil(512)
 /// ```
 /// This ensures that tools like `du` and `ls -s` report reasonable values.
-pub const FUSE_BLOCK_SIZE: u32 = 512;
+pub const FUSE_BLOCK_SIZE: u64 = 512;
 
 /// Synthesizes a `FileAttr` structure for a given inode and size.
 ///
@@ -184,7 +184,7 @@ pub fn make_attr(ino: u64, size: u64, uid: u32, gid: u32) -> FileAttr {
     FileAttr {
         ino,
         size,
-        blocks: size.div_ceil(FUSE_BLOCK_SIZE as u64),
+        blocks: size.div_ceil(FUSE_BLOCK_SIZE),
         atime: UNIX_EPOCH,
         mtime: UNIX_EPOCH,
         ctime: UNIX_EPOCH,
@@ -204,6 +204,6 @@ pub fn make_attr(ino: u64, size: u64, uid: u32, gid: u32) -> FileAttr {
         gid,
         rdev: 0,
         flags: 0,
-        blksize: FUSE_BLOCK_SIZE,
+        blksize: FUSE_BLOCK_SIZE as u32,
     }
 }
